requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
CacheControl==0.14.4
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
curl_cffi==0.15.0
dataclasses-json==0.6.7
distro==1.9.0
filetype==1.2.0
finnhub-python==2.4.27
firebase_admin==7.3.0
frozendict==2.4.7
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-cloud-core==2.5.1
google-cloud-firestore==2.26.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-genai==1.71.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.15
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.28
langchain-google-genai==4.2.1
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.27
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.1
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.19.0
newsapi-python==0.2.7
numpy==2.4.4
openai==2.31.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.2
peewee==4.0.4
pillow==12.2.0
platformdirs==4.9.4
plotly==6.6.0
portalocker==3.2.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.20.0
PyJWT==2.12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
qdrant-client==1.17.1
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
streamlit==1.56.0
tenacity==9.1.4
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
yfinance==1.2.1
zstandard==0.25.0
